rename hook for live session slide state
